+++ src/lib/utils.js
@@ -41,3 +41,10 @@ export function arrayEquals (a, b) {
 
   return true;
 }
+
+export function addErrorType (type) {
+  return (err) => {
+    err.type = type;
+    throw err;
+  };
+}